move get_binary_in_path to analyzer module
this function is only used in the analyzer package no need for it
to be in the codechecker_common.util
