+++ analyzer/codechecker_analyzer/analyzers/clangtidy/analyzer.py
@@ -15,7 +15,6 @@ import shlex
 import subprocess
 
 from codechecker_common.logger import get_logger
-from codechecker_common.util import get_binary_in_path
 
 from codechecker_analyzer import host_check
 from codechecker_analyzer import env
@@ -207,9 +206,9 @@ class ClangTidy(analyzer_base.SourceAnalyzer):
             return False
 
         # clang-tidy, clang-tidy-5.0, ...
-        clangtidy = get_binary_in_path(['clang-tidy'],
-                                       r'^clang-tidy(-\d+(\.\d+){0,2})?$',
-                                       environ)
+        clangtidy = env.get_binary_in_path(['clang-tidy'],
+                                           r'^clang-tidy(-\d+(\.\d+){0,2})?$',
+                                           environ)
 
         if clangtidy:
             LOG.debug("Using '%s' for Clang-tidy!", clangtidy)
